website/pyproject.toml
[project]
name = "website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fpdf>=1.7.2",
    "googlesearch-python>=1.3.0",
    "ollama>=0.5.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-magic>=0.4.27",
    "reportlab>=4.4.1",
    "streamlit>=1.45.1",
]
